Update depends on file in migration
This ext depends on changes made after beta4 was released. It depends on changes that came with the ACP Contact Admin form, so I'm changing the 1st migration to depend on that change's migration. Maybe this should be changed again to beta5 later on, but for now this could prevent installs on out-dated versions of phpBB (which have already occurred).